Use KActionMenu instead of KToolBarPopupAction for the "New" and "Synchronize" menus that are built on the fly.
Just gives a one consistent feel for the toolbar, I think.
KToolBarPopupAction is "... designed for history actions
and for any other action that has more detail in a toolbar
than in a menu (e.g. tool chooser with ’Other’ leading to
a dialog...)."  So I think KActionMenu is more appropriate.

I know there was a bug report about this, but I can't find it.


svn path=/trunk/KDE/kdepim/kontact/src/; revision=956261