Backport r959710 by smartins from trunk to the 4.2 branch:
If the user used RMB in the blank space (outside all item) the context menu was acting upon the previously selected item. If the user
had for example an akonadi resource selected it was impossible to add another top level resource without restarting korg, it
was only possible to add subresources.

Now, RMB in the blank space will ignore selected items and always show the generic add resource dialog.

svn path=/branches/KDE/4.2/kdepim/; revision=960069