Backport r960620 by jtamate from trunk to the 4.2 branch:
CCBUG: 88271
CCBUG: 152707

Use the right regular expresion to check for mail headers.
Also, use a mini ad-hoc state machine to not treat all that looks like headers as headers.
When a blank line is found, it is supposed to start a body part, until a header with Content- is found.

svn path=/branches/KDE/4.2/kdepim/; revision=961651