Backport r965847 by smartins from trunk to the 4.2 branch:
Fix broken signal connection.

KDialog's slotDelayedDestruct() is named delayedDestruct() in KDE4.

This means that now, if you have your attachment editor open, and minimize/restore korganizer the attachment dialog will be gone, (as
hidden() is connected to delayedDestruct()).

Not sure if this feature is very useful, but that's what we had in KDE3 and wasn't working in KDE4 because slotDelayedDestruct
doesn't exist anywhere.

svn path=/branches/KDE/4.2/kdepim/; revision=966699