Backport r968878 by smartins from trunk to the 4.2 branch:
baseview.h declares the virtual function selectedDates() and says it should return the selected incidence's date.

Monthview's implementation is returning the selected cell's date, not the incidence's. So, if you had a recurring event and
tried to delete an occurrence then the selected cell's date would be used while deleting.

svn path=/branches/KDE/4.2/kdepim/; revision=968885