Reminders on to-dos didn't work most of the time because they were relative do start by default and many to-dos only have due-date.
Now, when you check the reminder box, if the to-do has a due date, it will be relative to end, if it only has start date it will be 
relative to start ( not allowed by UI yet ).

I'm not closing this bug yet, we should protect the advanced alarm editor so the user can't choose invalid "relative to" values, new 
strings needed.


CCBUG: 159822


svn path=/trunk/KDE/kdepim/; revision=975902