--hard_coded_colors
This removes the use of hard-coded colors as defaults for the alarm list and default alarm message. (This isn't all occurrences, but the most obvious.) David and I agreed on the choice of color roles being okay after the default VisitedText color was made darker to better contrast with InactiveText.

REVIEW: 1288

svn path=/trunk/KDE/kdepim/; revision=1013077