Remove the Tracker integration from the Akonadi port of KMail.
It needs to be ported to Akonadi anyway, but the more important reason is that
the Tracker integration can and should now be done with a seperate application,
which the design of Akonadi now allows.

This will only affect the Akonadi-based KMail, which is expected to be released
with KMail 2.

CCMAIL: philip@codeminded.be

svn path=/branches/work/akonadi-ports/kdepim/; revision=1066789