Add a configuration option "Open links in new tab instead of in new window", the same as in Konqueror.
Honour link TARGET="_blank" in the article viewer and browser tabs,
open a new window or a new tab in accordance with that option.

Open a new tab in the background, if the user prefers background tabs
(i.e. if they have configured that for the left mouse button).

Reviewed at http://reviewboard.kde.org/r/3008

svn path=/trunk/KDE/kdepim/akregator/; revision=1097942