libkleopatrarc: make argument passing more robust for Archive and Checksum Definitions
This requires GNU-compatible tools (for -- and tar --null), and a POSIX system (for xargs),
so it won't work on Windows out of the box (it never did, anyway, unless you had the tools
in PATH). ZIP cannot be called safely, since it accepts options after positional parameters,
doesn't support --, and can accept filenames on stdin only in a newline-separated way (which
breaks when filenames contain newlines).

We might need a way to specify which platform a given Archive Definition / Checksum Definition
is compatible with...

svn path=/branches/kdepim/enterprise4/kdepim/; revision=1118267