libkpimidentities: add email aliases to Identity; use them in KMail, KOrg, wizards/
With this patch, KMail allows an identity to have arbitrarily many email 
aliases. Aliases are email addresses that are considered representing the 
identity, but that will not be used for outgoing messages.

The immediate problem this is fixing is that hitting 'A' on a message with 
an old (=alias) address includes that (old/alias) address in the list of To's.

svn path=/branches/kdepim/enterprise/kdepim/; revision=1122656