Adapted history/undo to the async nature of akonadi.
Splited undo() into undo() and undoFinished(), the last one is called when the job finishes
so it can make redo available. ( redo was being made available before the job ended ).

Splited redo() into redo() and redoFinished() (same as above).

Fixed problem where it was trying to delete a non existant item, that happens
because when you "undo a delete" or "redo an add" a new Akonadi::Item is created and it
has a new Id, so we must listen to IncidenceChanger's signal and update all entries with
the new id.

svn path=/trunk/KDE/kdepim/; revision=1128933