Created function: bool IncidenceChanger::canDelete( Akonadi::Item::Id id )
Returns false if there is or there was a running delete job for that item id.

Will solve problems where the user presses "Del" twice and that triggers two delete jobs, one of which will result in an error.

This is also better than asking the model because the ETM will still hold the item for some milisecs after the delete job ends.


svn path=/trunk/KDE/kdepim/akonadi/kcal/; revision=1130468