Make the recurrence dialog a real modal dialog.
In the desktop version advanced recurrence editting is done in a separate
dialog. This was once extracted from a tab page in the incidence editor into
this dialog. Therefore it was behaving quite strange in comparison to other
dialogs. Currently it still makes use of the IncidenceRecurrenceEditor but this
one isn't combined with the date time editor any longer. For the mobile variant
we might need to do things slightly different. That's not sorted out yet.

svn path=/branches/work/komo/kdepim/; revision=1136729