Backport r1137804 by smartins from trunk to e35:
Only archive a completed to-do if it's children are also complete, otherwise we get a crash because of orphan to-dos.

Archive now has the same behaviour as purge: purge refused to make children orphan.

Won't svnmerge record this commit because it has a new string that should be fwd ported.


svn path=/branches/kdepim/enterprise/kdepim/; revision=1137845