attendee editor: qcombobox wasn't extensible enough, so add this new widget that acts like a combobox but looks like a qtoolbutton.
The goal here is to have a combobox like dropdown for selecting items with text+icon, but what is normally the combolabel should be just an icon in order to cut down on horizontal space.

This widget pretends to act like a combobox by implementing addItem(s) and (set)CurrentIndex, but underneath it is really storing the items as actions in a QMenu.

Very experimental.

svn path=/branches/work/komo/kdepim/; revision=1141947