Try to improve kmail-mobile performance again
To achieve the best performance for mobile applications using
QGV, the best way right now is to use the GLWidget as the viewport
(so it will render what is possible using opengl) and use raster as
the graphicssystem. We should use raster because the opengl graphics system
is not that good right now and needs improvements.

svn path=/trunk/KDE/kdepim/mobile/; revision=1149605