Avoid initializing the tags list from nepomuk for sizehint calculation.
Since the font for each message can be overridden by a tag, and since
the font size needs to be known at row height calculation time, the
tags list ended up being populated for each item, during message list
creation, which, since it's a nepomuk roundtrip or two, was very slow.
This change makes the font only take the tag into account when the tags
list is already populated. Not perfect, as for non visible items the
height might be off, if they have font override tags, but I guess we
can live with the consequences. It's rather a fringe feature anyhow.

CCBUG: 244503

svn path=/branches/KDE/4.5/kdepim/; revision=1150006