Composer edit 'auto-size' fix
We should also set the maximum height in order to make it shrink
when text is removed from the composer. The initial size of the
widget is a good guess of it's minimum size to fit the current
screen as it was already layouted. Sad that QML doesn't have sizeHints
and had to include this work around.

svn path=/trunk/KDE/kdepim/mobile/; revision=1155613