Make the flickable area follow the cursor
This commit fixes some old code that was presuming that we were using
a QAbstractKineticScroller. Now it works with the QML Flickable area
and also on the desktop (no specific Maemo5 code anymore).

It loops on all the parents of the ProxyWidget, looking for a
QDeclarativeItem that has the property 'flickableData' that right now
just the QDeclarativeFlickable item has. It's not the best way but it's
the only way to find the item before this class becomes a public class.

Reviewed-by: Anselmo L. de Melo <anselmo.melo@openbossa.org>

svn path=/trunk/KDE/kdepim/mobile/; revision=1159259