Make strings properly translatable.
i18n( "the %1" ).arg( foo ), with foo being "task" or "invitation", can
not be properly translated in all languages, as the definite article
needs to change depending on the following noun.

Actually not a problem for German this time, since "task" and "invitation"
happen to have the same article, but this might be different in other
languages.


svn path=/branches/kdepim/enterprise/kdepim/; revision=1159408