Fwd port r1170331 from 4.5:
When an item is resized, the itemModified( AgendaItem ) signal is emitted.
When resizing a multi-day event ( has more than one AgendaItem ), the item that must go with the signal is the first one,
otherwise alll hell breaks loose inside KOAgendaView::updateEventDates().
updateEventDates() is already complicated enough and deals with lots of corner cases, so i wont adapt it to support
AgendaItems other than the first. Fix the signal emitter instead.

(Already svnmerge recorded.)

svn path=/trunk/KDE/kdepim/calendarviews/; revision=1170340